backend/foodgram_backend/requirements.txt
Django==5.1.4
django-import-export==4.3.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
pillow==11.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
drf_extra_fields==3.7.0
django-filter==23.1